integration with VIS

+++ lib/visualizer.py
@@ -15,12 +15,13 @@ Modified:
 """
 import os
 import shutil
+from utils.dataWorker import dataWorker
 import json
 import requests as req
-import matplotlib as mpl
-if os.environ.get('DISPLAY','') == '':
-    print('No display found; using non-interactive Agg backend...\n')
-    mpl.use('Agg')
+#import matplotlib as mpl
+#if os.environ.get('DISPLAY','') == '':
+#    print('No display found; using non-interactive Agg backend...\n')
+#    mpl.use('Agg')
 
 
 class Visualizer():
@@ -31,7 +32,7 @@ class Visualizer():
     software.
     """
 
-    def __init__(self, config):
+    def __init__(self, config, log=None, rank=-1):
         """This is the constructor for the Visualizer class.
         The visualizer has two main modes defined by the `VizMethod` variable.
             - offline: writes .json files that the visualization software parses
@@ -40,7 +41,10 @@ class Visualizer():
         Args:
             config: config object
         """
-        self.vizMethod = config['Visualizer']['VizMethod']
+        self.log = log
+        self.rank = rank
+
+        self.vizMethod = str(config['Visualizer']['VizMethod'])
         self.vizUrl = config['Visualizer']['VizUrl']
         self.outputDir = config['Visualizer']['OutputDir']
         #self.traceHeader = {'data': 'trace'}
@@ -49,20 +53,39 @@ class Visualizer():
         self.commData = []
         self.anomalyData = []
 
-        if self.vizMethod == 'offline' and os.path.exists(self.outputDir):
-            shutil.rmtree(self.outputDir)
+        if self.rank >= 0:
+            self.outputDir = self.outputDir + '-{:03d}'.format(self.rank)
 
-        if self.vizMethod == 'offline' and not os.path.exists(self.outputDir):
-            os.makedirs(self.outputDir)
+        # potentinally, there could be multiple workers depending on the performance.
+        # currently, it only creates single worker.
+        self.worker = None
+        if self.vizMethod in ['online', 'offline']:
+            self.worker = None
+            n_workers = int(config['Visualizer']['UseWorker'])
+            if n_workers > 0:
+                self.worker = dataWorker(log=log)
 
+            if self.vizMethod == 'offline' and os.path.exists(self.outputDir):
+                shutil.rmtree(self.outputDir)
+
+            if self.vizMethod == 'offline' and not os.path.exists(self.outputDir):
+                os.makedirs(self.outputDir)
+
+    def join(self, force_to_quit):
+        if self.worker is not None:
+            self.worker.join(force_to_quit, self.rank)
 
     def sendReset(self):
         """Send signal to visualization server to restart itself."""
+        if self.vizMethod == 'off': return
+
         resetDict = {'type': 'reset'}
         if self.vizMethod == "online":
-            req.post(self.vizUrl, json=resetDict)
+            req.post(self.vizUrl + '/events', json=resetDict)
         elif self.vizMethod == "offline":
-            fn = "reset.json"
+            fn = "reset.{:03d}.json".format(self.rank) \
+                if self.rank >= 0 \
+                else "reset.json"
             with open(os.path.join(self.outputDir, fn), 'w') as outfile:
                 json.dump(resetDict, outfile, indent=4, sort_keys=True)
         else:
@@ -76,11 +99,16 @@ class Visualizer():
                 such as function EXIT/ENTRY or SEND/RECEIVE for communication events.
             frame_id (int): frame id
         """
+        if self.vizMethod == 'off': return
+
         eventDict = {'type': 'event_types', 'value': eventType}
         if self.vizMethod == "online":
-            r = req.post(self.vizUrl, json=eventDict)
-        if self.vizMethod == "offline":
-            fn = "et.{:06d}.json".format(frame_id)
+            r = req.post(self.vizUrl + '/events', json=eventDict)
+        elif self.vizMethod == "offline":
+            fn = "et.{:06d}.{:03d}.json".format(frame_id, self.rank) \
+                if self.rank >= 0 \
+                else "et.{:06d}.json".format(frame_id)
+
             with open(os.path.join(self.outputDir, fn), 'w') as outfile:
                 json.dump(eventDict, outfile)
         else:
@@ -95,17 +123,21 @@ class Visualizer():
             funMap (dictionary): function id (int) key and function name (string) value
             frame_id (int): frame id
         """
+        if self.vizMethod == 'off': return
+
         funDict = {'type': 'functions', 'value': funMap}
         if self.vizMethod == "online":
             req.post(self.vizUrl, json={'type': 'functions', 'value': funMap})
         elif self.vizMethod == "offline":
-            fn = "functions.{:06d}.json".format(frame_id)
+            fn = "functions.{:06d}.{:03d}.json".format(frame_id, self.rank) \
+                if self.rank >=0 \
+                else "functions.{:06d}.json".format(frame_id)
             with open(os.path.join(self.outputDir, fn), 'w') as outfile:
                 json.dump(funDict, outfile, indent=4, sort_keys=True)
         else:
             raise ValueError("Unsupported method: %s" % self.vizMethod)
 
-    def sendData(self,
+    def sendData_v1(self,
                  funData, countData, commData, funOfInt, outlId, frame_id=0,
                  eventType=None, funMap=None):
         """
@@ -121,10 +153,11 @@ class Visualizer():
             frame_id: (int), frame id
             eventType: (list), event types
         """
-        #todo: is this correct way to aggregate all data
+        if self.vizMethod == 'off': return
+
         dataList = []
         dataList += funData
-        dataList += countData
+        #dataList += countData
         dataList += commData
 
         traceDict = {
@@ -142,19 +175,102 @@ class Visualizer():
             traceDict['value']['functions'] = funMap
 
         if self.vizMethod == "online":
-            req.post(self.vizUrl, json=traceDict)
+            if self.worker is not None:
+                self.worker.put(self.vizMethod, self.vizUrl + '/events', traceDict,
+                                self.rank, frame_id)
+            else:
+                try:
+                    r = req.post(self.vizUrl + '/events', json=traceDict)
+                    r.raise_for_status()
+                except req.exceptions.HTTPError as e:
+                    if self.log is not None: self.log.info("Http Error: ", e)
+                except req.exceptions.ConnectionError as e:
+                    if self.log is not None: self.log.info("Connection Error: ", e)
+                except req.exceptions.Timeout as e:
+                    if self.log is not None: self.log.info("Timeout Error: ", e)
+                except req.exceptions.RequestException as e:
+                    if self.log is not None: self.log.info("OOps: something else: ", e)
+                except Exception as e:
+                    if self.log is not None: self.log.info("Really unknown error: ", e)
+
         elif self.vizMethod == "offline":
-            fn = "trace.{:06d}.json".format(frame_id)
-            with open(os.path.join(self.outputDir, fn), 'w') as outfile:
-                json.dump(traceDict, outfile, indent=4, sort_keys=True)
+            fn = "trace.{:06d}.{:03d}.json".format(frame_id, self.rank) \
+                if self.rank>=0 \
+                else "trace.{:06d}.json".format(frame_id)
+            if self.worker is not None:
+                self.worker.put(self.vizMethod, os.path.join(self.outputDir, fn), traceDict,
+                                self.rank, frame_id)
+            else:
+                with open(os.path.join(self.outputDir, fn), 'w') as outfile:
+                    json.dump(traceDict, outfile, indent=4, sort_keys=True)
+
         else:
             raise ValueError("Unsupported method: %s" % self.vizMethod)
-            # This line of code was added to check whether we ever encounter a non nan value in countData[i][7]
-            #===================================================================
-            # dataList = funData.tolist() + countData.tolist() + commData.tolist()
-            # for i in range(0,countData.shape[0]):
-            #     if str(countData[i][7]) != str('nan'):
-            #         print(type(countData[i][7]))
-            #         print(str(countData[i][7]))
-            #         raise Exception("countData has non nan entry in column 7")
-            #===================================================================
+
+    def sendData_v2(self, execData, funMap, getStat, frame_id=0):
+        """
+        Send function, counter and communication data as well as the results of the analysis
+        of interest and id(s) of outlier data points.
+
+        Args:
+        """
+        if self.vizMethod == 'off': return
+
+        execDict = {}
+        stat = {}
+
+        for funid, execList in execData.items():
+            for d in execList:
+                if d.label == 1:
+                    continue
+                key = d.get_id()
+                value = d.to_dict()
+                execDict[str(key)] = value
+
+            n_total, n_abnormal, mean, std = getStat(funid)
+            key = funMap[funid]
+            stat[key] = {
+                "abnormal": n_abnormal,
+                "regular": n_total - n_abnormal,
+                "mean": mean,
+                "std": std,
+            }
+
+        traceDict = {
+            'executions': execDict,
+            'stat': stat
+        }
+
+
+        if self.vizMethod == "online":
+            if self.worker is not None:
+                self.worker.put(self.vizMethod, self.vizUrl + '/executions', traceDict,
+                                self.rank, frame_id)
+            else:
+                try:
+                    r = req.post(self.vizUrl + '/executions', json=traceDict)
+                    r.raise_for_status()
+                except req.exceptions.HTTPError as e:
+                    if self.log is not None: self.log.info("Http Error: ", e)
+                except req.exceptions.ConnectionError as e:
+                    if self.log is not None: self.log.info("Connection Error: ", e)
+                except req.exceptions.Timeout as e:
+                    if self.log is not None: self.log.info("Timeout Error: ", e)
+                except req.exceptions.RequestException as e:
+                    if self.log is not None: self.log.info("OOps: something else: ", e)
+                except Exception as e:
+                    if self.log is not None: self.log.info("Really unknown error: ", e)
+
+        elif self.vizMethod == "offline":
+            fn = "trace.{:06d}.{:03d}.json".format(frame_id, self.rank) \
+                if self.rank >= 0 \
+                else "trace.{:06d}.json".format(frame_id)
+            if self.worker is not None:
+                self.worker.put(self.vizMethod, os.path.join(self.outputDir, fn), traceDict,
+                                self.rank, frame_id)
+            else:
+                with open(os.path.join(self.outputDir, fn), 'w') as outfile:
+                    json.dump(traceDict, outfile, indent=4, sort_keys=True)
+
+        else:
+            raise ValueError("Unsupported method: %s" % self.vizMethod)
